schema builder refactoring dtos
